match-bot/pyproject.toml
[tool.poetry]
name = "match-bot"
version = "0.1.0"
description = ""
authors = ["CoSm1C AvEnGeR <zhora.zhilin.06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.28"
asyncpg = "0.29.0"
redis = "4.6.0"
aiokafka = "^0.10.0"
aiogram = "3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
